remove architecture.py to use core file
