+++ kaolin/models/Pixel2Mesh.py
@@ -38,65 +38,62 @@ class VGG(nn.Module):
               year         = "2015",
             }
     """
-    
-    def __init__(self, channels = 4):
-        
+    def __init__(self, channels=4):
         super(VGG, self).__init__()
 
-        self.layer0_1 = nn.Conv2d(channels, 16, 3, stride = 1, padding = 1)
-        self.layer0_2 = nn.Conv2d(16, 16, 3, stride = 1, padding = 1)
-        
-        self.layer1_1 = nn.Conv2d(16, 32, 3, stride = 2, padding = 1) 
-        self.layer1_2 = nn.Conv2d(32, 32, 3, stride = 1, padding = 1)
-        self.layer1_3 = nn.Conv2d(32, 32, 3, stride = 1, padding = 1)
-        
-        self.layer2_1 = nn.Conv2d(32, 64, 3, stride = 2, padding = 1) 
-        self.layer2_2 = nn.Conv2d(64, 64, 3, stride = 1, padding = 1)
-        self.layer2_3 = nn.Conv2d(64, 64, 3, stride = 1, padding = 1)
-        
-        self.layer3_1 = nn.Conv2d(64, 128, 3, stride = 2, padding = 1) 
-        self.layer3_2 = nn.Conv2d(128, 128, 3, stride = 1, padding = 1)
-        self.layer3_3 = nn.Conv2d(128, 128, 3, stride = 1, padding = 1)
-        
-        self.layer4_1 = nn.Conv2d(128, 256, 5, stride = 2, padding = 2) 
-        self.layer4_2 = nn.Conv2d(256, 256, 3, stride = 1, padding = 1)
-        self.layer4_3 = nn.Conv2d(256, 256, 3, stride = 1, padding = 1)
-        
-        self.layer5_1 = nn.Conv2d(256, 512, 5, stride = 2, padding = 2) 
-        self.layer5_2 = nn.Conv2d(512, 512, 3, stride = 1, padding = 1)
-        self.layer5_3 = nn.Conv2d(512, 512, 3, stride = 1, padding = 1)
-        self.layer5_4 = nn.Conv2d(512, 512, 3, stride = 1, padding = 1)
-        
+        self.layer0_1 = nn.Conv2d(channels, 16, 3, stride=1, padding=1)
+        self.layer0_2 = nn.Conv2d(16, 16, 3, stride=1, padding=1)
+
+        self.layer1_1 = nn.Conv2d(16, 32, 3, stride=2, padding=1)
+        self.layer1_2 = nn.Conv2d(32, 32, 3, stride=1, padding=1)
+        self.layer1_3 = nn.Conv2d(32, 32, 3, stride=1, padding=1)
+
+        self.layer2_1 = nn.Conv2d(32, 64, 3, stride=2, padding=1)
+        self.layer2_2 = nn.Conv2d(64, 64, 3, stride=1, padding=1)
+        self.layer2_3 = nn.Conv2d(64, 64, 3, stride=1, padding=1)
+
+        self.layer3_1 = nn.Conv2d(64, 128, 3, stride=2, padding=1)
+        self.layer3_2 = nn.Conv2d(128, 128, 3, stride=1, padding=1)
+        self.layer3_3 = nn.Conv2d(128, 128, 3, stride=1, padding=1)
+
+        self.layer4_1 = nn.Conv2d(128, 256, 5, stride=2, padding=2)
+        self.layer4_2 = nn.Conv2d(256, 256, 3, stride=1, padding=1)
+        self.layer4_3 = nn.Conv2d(256, 256, 3, stride=1, padding=1)
+
+        self.layer5_1 = nn.Conv2d(256, 512, 5, stride=2, padding=2)
+        self.layer5_2 = nn.Conv2d(512, 512, 3, stride=1, padding=1)
+        self.layer5_3 = nn.Conv2d(512, 512, 3, stride=1, padding=1)
+        self.layer5_4 = nn.Conv2d(512, 512, 3, stride=1, padding=1)
+
     def forward(self, img):
-        
         img = F.relu(self.layer0_1(img))
         img = F.relu(self.layer0_2(img))
-        
+
         img = F.relu(self.layer1_1(img))
         img = F.relu(self.layer1_2(img))
         img = F.relu(self.layer1_3(img))
-        
+
         img = F.relu(self.layer2_1(img))
         img = F.relu(self.layer2_2(img))
         img = F.relu(self.layer2_3(img))
-        A = torch.squeeze(img) 
-        
+        A = torch.squeeze(img)
+
         img = F.relu(self.layer3_1(img))
         img = F.relu(self.layer3_2(img))
         img = F.relu(self.layer3_3(img))
-        B = torch.squeeze(img) 
-        
+        B = torch.squeeze(img)
+
         img = F.relu(self.layer4_1(img))
         img = F.relu(self.layer4_2(img))
         img = F.relu(self.layer4_3(img))
-        C = torch.squeeze(img) 
-        
+        C = torch.squeeze(img)
+
         img = F.relu(self.layer5_1(img))
         img = F.relu(self.layer5_2(img))
         img = F.relu(self.layer5_3(img))
         img = F.relu(self.layer5_4(img))
-        D = torch.squeeze(img) 
-        
+        D = torch.squeeze(img)
+
         return [A, B, C, D]
 
 
@@ -116,7 +113,6 @@ class G_Res_Net(nn.Module):
               year={2018}
             }
     """
-
     def __init__(self, input_features, hidden = 128, output_features = 3):
         super(G_Res_Net, self).__init__()
         self.gc1 = GCN(input_features, hidden)
@@ -214,7 +210,6 @@ class GCN(nn.Module):
             self.bias.data.uniform_(-.1, .1)
 
     def forward(self, input, adj):
-
         support = torch.bmm(input, self.weight.unsqueeze(0).expand(input.shape[0], -1, -1))
 
         output = torch.bmm(adj.unsqueeze(0).expand(input.shape[0], -1, -1), support)
@@ -223,6 +218,4 @@ class GCN(nn.Module):
         return output
 
     def __repr__(self):
-        return self.__class__.__name__ + ' (' \
-               + str(self.in_features) + ' -> ' \
-               + str(self.out_features) + ')'
+        return self.__class__.__name__ + ' (' + str(self.in_features) + ' -> ' + str(self.out_features) + ')'
